Improve styles
This commit improves a lot of the styles on the landings index page.

Buttons are added to each group that allow you to add an item to that
group. It defaults to it's own group, but that is easy to change on the
form.

There are edit and remove icons for each item, and editing the group
will be added in future commits.

The image background is set to cover the entire view, instead of
tiling.