Change JS to IIFY, handle destroy errors, JS still broken
This commit changes the JS to the IIFY style in an attempt to get it to
work, but it's still not loading the javascript when I load the edit
page. When I reload it works.

Added a branch for the destroy action to display an error if a group
couldn't be deleted for some reason.